Remove query_engine.hpp's dependency on query_expression.hpp.
The single type that query_engine.hpp needs is moved to a new
query_operators.hpp. This now avoids rebuilding all non-query .cpp files
when query_expression.hpp changes.
